Move controllers to subpackages, change an Integration Test to a Unit Test
