+++ src/main/java/com/example/springboot/templateControllers/DepartmentController.java
@@ -2,6 +2,7 @@ package com.example.springboot.templateControllers;
 
 import com.example.springboot.entities.Department;
 import com.example.springboot.repositories.DepartmentRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,7 +11,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 public class DepartmentController {
     private final DepartmentRepository departmentRepository;
 
-    public DepartmentController(DepartmentRepository departmentRepository) {
+    public DepartmentController(@Autowired DepartmentRepository departmentRepository) {
         this.departmentRepository = departmentRepository;
     }
