+++ src/test/resources/application.properties
@@ -1,8 +1,8 @@
 endpoints.sensitive=false
 management.endpoints.web.exposure.include=*
 debug=true
-
 # https://www.baeldung.com/spring-testing-separate-data-source
+spring.jpa.hibernate.ddl-auto=create-drop
 spring.datasource.driver-class-name=org.h2.Driver
 spring.datasource.url=jdbc:h2:mem:db;DB_CLOSE_DELAY=-1
 spring.datasource.username=sa
